Time: 0 ms (100.00%), Space: 7.4 MB (70.13%) - LeetHub

+++ 818-race-car/818-race-car.cpp
@@ -2,17 +2,17 @@ class Solution {
 public:
     
     int racecar(long int target) {
-        queue<pair<long int, long int>> q;
+        queue<pair<int, long int>> q;
         q.push({0, 1});
         long int depth = 0;
         
         while (true) {
             //vector<pair<long int, long int>> next_q;
             int n = q.size(); 
-            for (long int idx = 0; idx < n; idx++) {
-                pair<long int, long int> node = q.front();
+            for (int idx = 0; idx < n; idx++) {
+                pair<int, long int> node = q.front();
                 q.pop();
-                long int pos = node.first;
+                int pos = node.first;
                 long int speed = node.second;
                 if (pos == target) {
                     return depth;
